python-tools/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
atproto==0.0.55
attrs==24.3.0
bitarray==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.7
cryptography==43.0.3
cytoolz==1.0.1
dnspython==2.7.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
libipld==3.0.0
morphys==1.0
multidict==6.1.0
parsimonious==0.10.0
propcache==0.2.1
py-multibase==1.0.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pysha3==1.0.2
python-baseconv==1.2.2
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
six==1.17.0
sniffio==1.3.1
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.6.1
websockets==13.1
yarl==1.18.3
